policy: Change Identity and Nexthdr types
Key Identity and Nexthdr fields are now defined as
identity.NumericIdentity and u8proto.U8proto, respectively. This improves
type safety and almost eliminates the need for type casts regarding
these.

+++ pkg/k8s/watchers/pod.go
@@ -971,7 +971,7 @@ func (k *K8sPodWatcher) updatePodHostData(oldPod, newPod *slim_corev1.Pod, oldPo
 			}
 			k8sMeta.NamedPorts[port.Name] = ciliumTypes.PortProto{
 				Port:  uint16(port.ContainerPort),
-				Proto: uint8(p),
+				Proto: p,
 			}
 		}
 	}
